Extract Residuals from Fast Regression Models

This function extracts residuals from a fast regression model table (fast_regression()).

Usage

extract_regression_residuals(.model_tbl)

Arguments

.model_tbl

A fast regression model specification table (fst_reg_spec_tbl).

Value

The function returns a list of data frames, each containing residuals, actual values, and predicted values for a specific model.

Details

The function checks if the input model specification table inherits the class 'fst_reg_spec_tbl' and if it contains the column 'pred_wflw'. It then manipulates the data, grouping it by model, and extracts residuals for each model. The result is a list of data frames, each containing residuals, actual values, and predicted values for a specific model.

See also

Author

Steven P. Sanderson II, MPH

Examples

library(recipes, quietly = TRUE)
library(dplyr, quietly = TRUE)

rec_obj <- recipe(mpg ~ ., data = mtcars)

fr_tbl <- fast_regression(mtcars, rec_obj, .parsnip_eng = c("lm","glm"),
.parsnip_fns = "linear_reg")
#> Error in dplyr::mutate(mod_fitted_tbl, pred_wflw = internal_make_wflw_predictions(mod_fitted_tbl,     splits_obj)): ℹ In argument: `pred_wflw =
#>   internal_make_wflw_predictions(mod_fitted_tbl, splits_obj)`.
#> Caused by error in `map2()`:
#> ℹ In index: 1.
#> Caused by error in `dplyr::filter()`:
#> ! object 'rec_obj' not found

result <- extract_regression_residuals(fr_tbl)
#> Error in extract_regression_residuals(fr_tbl): object 'fr_tbl' not found

head(result[[1]])
#> Error in head(result[[1]]): object 'result' not found