Internals Safely Make a Fitted Workflow from Model Spec tibble

Safely Make a fitted workflow from a model spec tibble.

Usage

internal_make_fitted_wflw(.model_tbl, .splits_obj)

Arguments

.model_tbl

The model table that is generated from a function like fast_regression_parsnip_spec_tbl(), must have a class of "tidyaml_mod_spec_tbl". This is meant to be used after the function internal_make_wflw() has been run and the tibble has been saved.

.splits_obj

The splits object from the auto_ml function. It is internal to the auto_ml_ function.

Value

A list object of workflows.

Details

Create a fitted parnsip model from a workflow object.

See also

Author

Steven P. Sanderson II, MPH

Examples

library(recipes, quietly = TRUE)

mod_spec_tbl <- fast_regression_parsnip_spec_tbl(
  .parsnip_eng = c("lm","glm"),
  .parsnip_fns = "linear_reg"
)

rec_obj <- recipe(mpg ~ ., data = mtcars)
splits_obj <- create_splits(mtcars, "initial_split")

mod_tbl <- mod_spec_tbl |>
  mutate(wflw = full_internal_make_wflw(mod_spec_tbl, rec_obj))

internal_make_fitted_wflw(mod_tbl, splits_obj)
#> [[1]]
#> ══ Workflow [trained] ══════════════════════════════════════════════════════════
#> Preprocessor: Recipe
#> Model: linear_reg()
#> 
#> ── Preprocessor ────────────────────────────────────────────────────────────────
#> 0 Recipe Steps
#> 
#> ── Model ───────────────────────────────────────────────────────────────────────
#> 
#> Call:
#> stats::lm(formula = ..y ~ ., data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp           hp         drat           wt  
#>   23.068535     0.167894     0.002459    -0.021997    -0.719376    -2.071675  
#>        qsec           vs           am         gear         carb  
#>    0.357921     2.749810     4.720170     0.567812    -1.235715  
#> 
#> 
#> [[2]]
#> ══ Workflow [trained] ══════════════════════════════════════════════════════════
#> Preprocessor: Recipe
#> Model: linear_reg()
#> 
#> ── Preprocessor ────────────────────────────────────────────────────────────────
#> 0 Recipe Steps
#> 
#> ── Model ───────────────────────────────────────────────────────────────────────
#> 
#> Call:  stats::glm(formula = ..y ~ ., family = stats::gaussian, data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp           hp         drat           wt  
#>   23.068535     0.167894     0.002459    -0.021997    -0.719376    -2.071675  
#>        qsec           vs           am         gear         carb  
#>    0.357921     2.749810     4.720170     0.567812    -1.235715  
#> 
#> Degrees of Freedom: 23 Total (i.e. Null);  13 Residual
#> Null Deviance:	    1002 
#> Residual Deviance: 71.63 	AIC: 118.4
#>