Internals Make Base Classification Tibble

Creates a base tibble to create parsnip classification model specifications.

Usage

make_classification_base_tbl()

Value

A tibble

Details

Creates a base tibble to create parsnip classification model specifications.

See also

Author

Steven P. Sanderson II, MPH

Examples

make_classification_base_tbl()
#> # A tibble: 31 × 3
#>    .parsnip_engine .parsnip_mode  .parsnip_fns       
#>    <chr>           <chr>          <chr>              
#>  1 earth           classification bag_mars           
#>  2 earth           classification discrim_flexible   
#>  3 dbarts          classification bart               
#>  4 MASS            classification discrim_linear     
#>  5 mda             classification discrim_linear     
#>  6 sda             classification discrim_linear     
#>  7 sparsediscrim   classification discrim_linear     
#>  8 MASS            classification discrim_quad       
#>  9 sparsediscrim   classification discrim_quad       
#> 10 klaR            classification discrim_regularized
#> # ℹ 21 more rows